计划以后还有别的iclientpy工程，参照 https://packaging.python.org/guides/packaging-namespace-packages/#pkgutil-style-namespace-packages 修改 __init.py
以下内容复制自网页以防链接失效：
pkgutil-style namespace packages
Python 2.3 introduced the pkgutil module and the extend_path function. This can be used to declare namespace packages that need to be compatible with both Python 2.3+ and Python 3. This is the recommended approach for the highest level of compatibility.

To create a pkgutil-style namespace package, you need to provide an __init__.py file for the namespace package:

setup.py
mynamespace/
    __init__.py  # Namespace package __init__.py
    subpackage_a/
        __init__.py  # Sub-package __init__.py
        module.py
The __init__.py file for the namespace package needs to contain only the following:

__path__ = __import__('pkgutil').extend_path(__path__, __name__)
Every distribution that uses the namespace package must include an identical __init__.py. If any distribution does not, it will cause the namespace logic to fail and the other sub-packages will not be importable. Any additional code in __init__.py will be inaccessible.

A complete working example of two pkgutil-style namespace packages can be found in the pkgutil namespace example project.